Move Incident Card
Moved IncidentCard to the page folder as it is only going to be used by that page.
